// nuxt-ui3/types/global.ts
export interface IdentityInfo {
  display: string
  email: string
  legal: string
  riot: string
  twitter: string
  web: string
}

export interface IIdentity {
  _id: string
  name: string
  info: IdentityInfo
  verified: boolean
  judgements: any[]
  parentIdentity: IIdentity
  sub: string
  subId: string
}

export interface ICandidateScore {
  address: string
  aggregate: number
  asn: number
  bonded: number
  councilStake: number
  country: number
  delegations: number
  democracy: number
  discovered: number
  faults: number
  inclusion: number
  location: number
  nominated: number
  nominatorStake: number
  offline: number
  provider: number
  randomness: number
  rank: number
  region: number
  spanInclusion: number
  total: number
  unclaimed: number
  updated: number
  __v?: number
  _id?: string
}

export interface ICandidateValidityItem {
  _id: string
  valid: boolean
  type: string
}

export interface ICandidate {
  chain?: string
  active: boolean
  commission: number
  councilStake: number
  councilVotes: number[]
  country: string
  democracyVoteCount: number
  democracyVotes: string[] | number[]
  discoveredAt: string | number //| Date
  faults: string[]
  stash: string
  total: number
  score: ICandidateScore
  rank: number
  identity: IIdentity
  location: string
  name: string
  nominatedAt: string | number //| Date
  onlineSince: number | string
  provider: string
  region: string
  rewardDestination?: string
  unclaimedEras: string[]
  valid: boolean
  validity: ICandidateValidityItem[]
  nominated_1kv: boolean
  nominations?: number | INomination[]
  version?: string
}

export interface INomination {}
// ...
export interface ICache {
  // eslint-disable-next-line
  items: Record<string, any>
}
// ...
interface IValidatorPrefs {
  commission: number
  blocked: boolean
}
export interface IValidator {
  active: boolean
  commission: number
  councilStake: number
  councilVotes: number[]
  democracyVoteCount: number
  // eslint-disable-next-line
  democracyVotes: any[]
  discoveredAt: number | Date
  // eslint-disable-next-line
  faults: any
  stash: string
  shortStash: string
  prefs: IValidatorPrefs
  nominators: string[]
  score: ICandidateScore
  rank: number
  identity: IIdentity
  location: string
  name: string
  nominatedAt: number | Date
  // eslint-disable-next-line
  onlineSince: any
  // eslint-disable-next-line
  total: number
  unclaimedEras: any[]
  valid: boolean
  // validity: ICandidateValidityItem[]
  StorageKey: number[]
  // eslint-disable-next-line
  Type: any
}